# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CartTask Api Document",
        "description": "This is the API documentation for cart task.",
        "contact": {
            "email": "alibabaeian670@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/product": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Index products",
                "description": "list of products",
                "operationId": "5cc57bec35d0f2045983cc8dcd00eccd",
                "responses": {
                    "200": {
                        "description": "List of products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Mac M2 Pro"
                                                    },
                                                    "price": {
                                                        "type": "integer",
                                                        "example": 50000
                                                    },
                                                    "stock": {
                                                        "type": "integer",
                                                        "example": 6
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Store product",
                "description": "Create a new product",
                "operationId": "1d0b0d8724685b7b8f9d87f403bfc4df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "price",
                                    "stock"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of the product",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "price": {
                                        "description": "Price of the product",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "stock": {
                                        "description": "Stock number",
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product has been created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "title": {
                                            "type": "string",
                                            "example": "Mac m2 pro"
                                        },
                                        "price": {
                                            "type": "number",
                                            "example": 100000000
                                        },
                                        "stock": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "text",
                                            "example": "Validation Failed"
                                        },
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "The price field is required."
                                                },
                                                "errors": {
                                                    "properties": {
                                                        "price": {
                                                            "type": "array",
                                                            "items": {
                                                                "type": "string",
                                                                "example": "The price field is required."
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}